feat: `util.env()` and `util.try_env()` (WIP)
Introduce utility functions for reading environment variables.

At the moment, the functions are available under the `Util` class of the new `util` module:

   bring util;
   log(util.Util.env("PATH");

In a subsequent PR we will "hoist" these static functions to become `util.env()`.

Additionally, we cannot currently test these inflight. I believe @yoav-steinberg is working to address this. Is that correct?

Resolves #2432

### Misc

Reorganize SDK non public utility functions. Most of them were actually only used by tests.
