+++ libs/wingsdk/package.json
@@ -41,7 +41,6 @@
   "devDependencies": {
     "@cdktf/provider-aws": "^12.0.1",
     "@types/aws-lambda": "^8.10.109",
-    "@types/debug": "^4.1.7",
     "@types/express": "^4.17.17",
     "@types/fs-extra": "^11.0.1",
     "@types/mime-types": "^2.1.1",
@@ -99,7 +98,6 @@
     "cdktf": "0.15.2",
     "constructs": "~10.1.228",
     "cron-parser": "^4.8.1",
-    "debug": "^4.3.4",
     "esbuild-wasm": "^0.17.4",
     "express": "^4.18.2",
     "ioredis": "^5.3.1",
@@ -128,7 +126,6 @@
     "aws-cdk-lib",
     "cdktf",
     "cron-parser",
-    "debug",
     "esbuild-wasm",
     "express",
     "ioredis",
