+++ libs/wingsdk/src/testing/simulator.ts
@@ -6,7 +6,7 @@ import { SDK_VERSION } from "../constants";
 import { ConstructTree } from "../core";
 // eslint-disable-next-line import/no-restricted-paths
 import { DefaultSimulatorFactory } from "../target-sim/factory.inflight";
-import { readJsonSync } from "../util";
+import { readJsonSync } from "../utils/misc";
 
 /**
  * Props for `Simulator`.
