+++ libs/wingsdk/src/util/util.ts
@@ -1,0 +1,25 @@
+/**
+ * Utility functions.
+ */
+export class Util {
+  /**
+   * Returns the value of an environment variable. Throws if not found or empty.
+   * @param name The name of the environment variable.
+   */
+  public static env(name: string): string {
+    const value = Util.tryEnv(name);
+    if (!value) {
+      throw new Error(`Environment variable ${name} not found.`);
+    }
+    return value;
+  }
+
+  /**
+   * Returns the value of an environment variable. Returns `nil` if not found or empty.
+   * @param name The name of the environment variable.
+   * @returns The value of the environment variable or `nil`.
+   */
+  public static tryEnv(name: string): string | undefined {
+    return process.env[name];
+  }
+}
