+++ libs/wingsdk/test/target-awscdk/bucket.test.ts
@@ -2,7 +2,7 @@ import { Match, Template } from "aws-cdk-lib/assertions";
 import { test, expect } from "vitest";
 import { Bucket } from "../../src/cloud";
 import * as awscdk from "../../src/target-awscdk";
-import { mkdtemp } from "../../src/util";
+import { mkdtemp } from "../util";
 
 const CDK_APP_OPTS = {
   stackName: "my-project",
