+++ libs/wingsdk/test/target-awscdk/queue.test.ts
@@ -4,7 +4,7 @@ import { Queue } from "../../src/cloud";
 import * as std from "../../src/std";
 import * as awscdk from "../../src/target-awscdk";
 import { Testing } from "../../src/testing";
-import { mkdtemp, sanitizeCode } from "../../src/util";
+import { mkdtemp, sanitizeCode } from "../util";
 
 const CDK_APP_OPTS = {
   stackName: "my-project",
