+++ libs/wingsdk/test/target-awscdk/topic.test.ts
@@ -1,10 +1,9 @@
-import { Match, Template } from "aws-cdk-lib/assertions";
+import { Template } from "aws-cdk-lib/assertions";
 import { test, expect } from "vitest";
 import { Topic } from "../../src/cloud";
-import * as std from "../../src/std";
 import * as awscdk from "../../src/target-awscdk";
 import { Testing } from "../../src/testing";
-import { mkdtemp, sanitizeCode } from "../../src/util";
+import { mkdtemp, sanitizeCode } from "../util";
 
 const CDK_APP_OPTS = {
   stackName: "my-project",
