+++ libs/wingsdk/test/target-sim/app.test.ts
@@ -3,7 +3,7 @@ import { test, expect } from "vitest";
 import { simulatorJsonOf } from "./util";
 import { Bucket } from "../../src/cloud";
 import * as sim from "../../src/target-sim";
-import { mkdtemp } from "../../src/util";
+import { mkdtemp } from "../util";
 
 test("app name can be customized", async () => {
   // GIVEN
