+++ libs/wingsdk/test/target-tf-aws/app.test.ts
@@ -3,7 +3,7 @@ import { test, expect } from "vitest";
 import { Function } from "../../src/cloud";
 import * as tfaws from "../../src/target-tf-aws";
 import { Testing } from "../../src/testing";
-import { mkdtemp } from "../../src/util";
+import { mkdtemp } from "../util";
 
 test("artifacts are located in app root level outdir", () => {
   // GIVEN
