+++ libs/wingsdk/test/target-tf-aws/logger.test.ts
@@ -2,8 +2,7 @@ import { test, expect } from "vitest";
 import { Function } from "../../src/cloud";
 import * as tfaws from "../../src/target-tf-aws";
 import { Testing } from "../../src/testing";
-import { mkdtemp, sanitizeCode } from "../../src/util";
-import { tfResourcesOf, tfSanitize } from "../util";
+import { mkdtemp, sanitizeCode, tfResourcesOf, tfSanitize } from "../util";
 
 test("inflight function uses a logger", () => {
   const app = new tfaws.App({ outdir: mkdtemp() });
