+++ libs/wingsdk/test/target-tf-aws/queue.test.ts
@@ -4,8 +4,13 @@ import { Queue } from "../../src/cloud";
 import * as std from "../../src/std";
 import * as tfaws from "../../src/target-tf-aws";
 import { Testing } from "../../src/testing";
-import { mkdtemp, sanitizeCode } from "../../src/util";
-import { tfResourcesOf, tfSanitize, treeJsonOf } from "../util";
+import {
+  mkdtemp,
+  sanitizeCode,
+  tfResourcesOf,
+  tfSanitize,
+  treeJsonOf,
+} from "../util";
 
 test("default queue behavior", () => {
   // GIVEN
