+++ libs/wingsdk/test/target-tf-azure/captures.test.ts
@@ -3,8 +3,7 @@ import { test, expect } from "vitest";
 import * as cloud from "../../src/cloud";
 import * as tfazure from "../../src/target-tf-azure";
 import { Testing } from "../../src/testing";
-import { mkdtemp, sanitizeCode } from "../../src/util";
-import { tfResourcesOf, tfSanitize } from "../util";
+import { mkdtemp, sanitizeCode, tfResourcesOf, tfSanitize } from "../util";
 
 test("function with a bucket binding requiring read_write", () => {
   // GIVEN
