+++ libs/wingsdk/test/testing/on-trace.test.ts
@@ -2,7 +2,7 @@ import { test, expect } from "vitest";
 import * as cloud from "../../src/cloud";
 import * as sim from "../../src/target-sim";
 import * as testing from "../../src/testing";
-import { mkdtemp } from "../../src/util";
+import { mkdtemp } from "../util";
 
 test("onTrace", async () => {
   // GIVEN
