+++ libs/wingsdk/test/testing/reload.test.ts
@@ -2,7 +2,7 @@ import { test, expect } from "vitest";
 import { Bucket } from "../../src/cloud";
 import * as sim from "../../src/target-sim";
 import * as testing from "../../src/testing";
-import { mkdtemp } from "../../src/util";
+import { mkdtemp } from "../util";
 
 test("reloading the simulator updates the state of the tree", async () => {
   let workdir = mkdtemp();
